ReadTextFromImages/requirements.txt
pytesseract
pillow
# sudo apt-get install tesseract-ocr